Moved step arguments class into it's own file
