+++ TestStack.BDDfy/TestStack.BDDfy.csproj
@@ -122,6 +122,7 @@
     </Compile>
     <Compile Include="Scanners\StepScanners\Fluent\IFluentScanner.cs" />
     <Compile Include="Scanners\StepScanners\Examples\WithExamplesExtensions.cs" />
+    <Compile Include="Scanners\StepScanners\Fluent\StepArgument.cs" />
     <Compile Include="Scenario.cs" />
     <Compile Include="Engine.cs" />
     <Compile Include="StepTitle.cs" />
